NULS extension - refactor Binance api
