+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/lisk/sources/binance/BinanceRateSource.java
@@ -73,15 +73,15 @@ public class BinanceRateSource implements IRateSource {
             return null;
         }
         final BinanceTickerData btcUsdt = api.getTicker("BTCUSDT");
-        final BinanceTickerData lskBtc = api.getTicker(cryptoCurrency + "BTC");
+        final BinanceTickerData selectedCryptoInBtc = api.getTicker(cryptoCurrency + "BTC");
         CoinmarketcapRateSource coinMarketCapSource = new CoinmarketcapRateSource(coinmarketcapApiKey, fiatCurrency);
         BigDecimal lastUsdtFiat = coinMarketCapSource.getExchangeRateLast("USDT", fiatCurrency);
-        if (lastUsdtFiat != null && btcUsdt.getPrice()!=null && lskBtc.getPrice() !=null ) {
+        if (lastUsdtFiat != null && btcUsdt.getPrice()!=null && selectedCryptoInBtc.getPrice() !=null ) {
             BigDecimal lastBtcPriceInUsdt = btcUsdt.getPrice();
-            BigDecimal lastLskPriceInBtc = lskBtc.getPrice();
-            BigDecimal lastLskPriceInUsdt = lastLskPriceInBtc.multiply(lastBtcPriceInUsdt);
-            BigDecimal lastLskPriceInFiat = lastLskPriceInUsdt.multiply(lastUsdtFiat);
-            return lastLskPriceInFiat;
+            BigDecimal lastSelectedCryptoPriceInBtc = selectedCryptoInBtc.getPrice();
+            BigDecimal lastSelectedCryptoPriceInUsdt = lastSelectedCryptoPriceInBtc.multiply(lastBtcPriceInUsdt);
+            BigDecimal lastSelectedCryptoPriceInFiat = lastSelectedCryptoPriceInUsdt.multiply(lastUsdtFiat);
+            return lastSelectedCryptoPriceInFiat;
         }
         return null;
     }
