+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/lisk/wallets/liskbinancewallet/BinanceSendCoinResponse.java
@@ -17,25 +17,22 @@
  ************************************************************************************/
 package com.generalbytes.batm.server.extensions.extra.lisk.wallets.liskbinancewallet;
 
-import java.math.BigDecimal;
-
-public class LskBinanceAssetData {
-    private String asset;
-    private BigDecimal free;
-
-    public String getAsset() {
-        return asset;
+public class BinanceSendCoinResponse {
+    private Boolean success; 
+    private String msg;
+    public Boolean getSuccess() {
+        return success;
     }
 
-    public void setAsset(String asset) {
-        this.asset = asset;
+    public void setSuccess(Boolean success) {
+        this.success = success;
+    }
+    
+    public String getMsg() {
+        return msg;
     }
 
-    public BigDecimal getFree() {
-        return free;
-    }
-
-    public void setFree(BigDecimal free) {
-        this.free = free;
+    public void setMsg(String msg) {
+        this.msg = msg;
     }
 }
