+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/lisk/wallets/liskbinancewallet/BinanceWalletAPI.java
@@ -23,13 +23,13 @@ import java.io.IOException;
 
 @Path("/")
 @Produces(MediaType.APPLICATION_JSON)
-public interface ILskBinanceAPI { 	
+public interface BinanceWalletAPI {
     @GET
     @Path("/api/v3/account")
-    LskBinanceResponse getCryptoBalance(@HeaderParam("X-MBX-APIKEY") String apiKey, @QueryParam("recvWindow") String recvWindow, @QueryParam("timestamp") String timeStamp, @QueryParam("signature") String signature ) throws IOException;
+    BinanceResponse getCryptoBalance(@HeaderParam("X-MBX-APIKEY") String apiKey, @QueryParam("recvWindow") String recvWindow, @QueryParam("timestamp") String timeStamp, @QueryParam("signature") String signature ) throws IOException;
 
     @POST
     @Path("wapi/v1/withdraw.html")
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
-    LskSendCoinResponse sendLsks(@HeaderParam("X-MBX-APIKEY") String apiKey, @QueryParam("asset") String asset, @QueryParam("address") String address, @QueryParam("amount") String amount, @QueryParam("name") String name, @QueryParam("recvWindow") String recvWindow, @QueryParam("timestamp") String timeStamp, @QueryParam("signature") String signature) throws IOException;
+    BinanceSendCoinResponse sendCryptoCurrency(@HeaderParam("X-MBX-APIKEY") String apiKey, @QueryParam("asset") String asset, @QueryParam("address") String address, @QueryParam("amount") String amount, @QueryParam("name") String name, @QueryParam("recvWindow") String recvWindow, @QueryParam("timestamp") String timeStamp, @QueryParam("signature") String signature) throws IOException;
 }
