+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/nuls/wallet/NulsWallet.java
@@ -1,12 +1,12 @@
 package com.generalbytes.batm.server.extensions.extra.nuls.wallet;
 
 import com.generalbytes.batm.common.currencies.CryptoCurrency;
-import com.generalbytes.batm.server.extensions.extra.lisk.wallets.liskbinancewallet.LskWallet;
+import com.generalbytes.batm.server.extensions.extra.lisk.wallets.liskbinancewallet.BinanceWallet;
 
 import java.util.HashSet;
 import java.util.Set;
 
-public class NulsWallet extends LskWallet {
+public class NulsWallet extends BinanceWallet {
 
     public NulsWallet(String address, String binanceApiKey, String binanceApiSecret) {
         super(address,binanceApiKey,binanceApiSecret);
